Fixed Otis being visibly cut off in CriminalCollapse
I knew about this bug since the start, but I didn't have the time. Widescreen resolutions would reveal the edges of the cutscene sprite. He is now always aligned with the right side of the screen. I also edited the first frame of the arm collection because the arm was cut off and it relatively useless to have.