Fixed Otis staying onscreen in TheyJumpinMe + consistency
Before, I used an animation to make the sprite empty, and then shortly after, would play another animation to animate Mr. Crimal. For some reason, I put both these animations in the same animator. These animations were separated by a 0.01 second delay, which is sometimes not slow enough for the computer. As a result, there was a chance both animations would occur in the same tick, and only the second one (the Crimal animation) would play.
The two animations are now functionally combined.
Regarding the "consistency" changes, the TextureRect nodes in TheyJumpinMe and ManImDead and that were named "ColorRect," I renamed to "TextureRect"