Made buttons pure white + removed font resource.
All the buttons on the pause menu and main menu have actually been a very light shade of gray. Fixed now. Also, I had created a separate unique font resource for the red "Are you sure?" text, because I thought I needed to in order to make it a different color. Removed now.