A handful of changes.
Made first mall train disappear when second spawns, restored the lack-of-collision for the water of the mall fountain, regrouped fountain water, maintained escalator animation, gave the escalator animation RESET tracks, added an exclamation point to the end of MONEYS in Steven's greet dialogue, and fixed a crash where somehow the TheyJumpinMe cutscene could load its manager script when running in the editor, but not in a build. As a result, it would just crash. This was caused from the capitalization renaming of script. Also, I might have made the mall dark by default again, but I can't remember if that was this commit or a previous one. I don't think these changes deserve separate commits. This will likely be the final commit before 1.1